Allow dev server to exit cleanly (SIGINT/SIGTERM)
Previously we were sending SIGKILL immediately to the child. This
ensures that the child exits as quickly as possible, but it also means
that the child probably won't have time to finish flushing files to
disk (trace files, cache hit statistics, etc.), and can lead to
incomplete writes.

We should give the child a small amount of time to exit cleanly with
SIGINT/SIGTERM before sending SIGKILL. The timeout is needed in case the
child process is unresponsive/hanging.